Version 0.05 beta

+++ src/WindowMngr.h
@@ -9,7 +9,7 @@ public:
 	static HWND CurrentWindow();
 	static std::wstring GetWindowList(tVariant* paParams, const long lSizeArray);
 	static std::wstring GetText(tVariant* paParams, const long lSizeArray);
-	static long GetWindowState(tVariant* paParams, const long lSizeArray);
+	static int32_t GetWindowState(tVariant* paParams, const long lSizeArray);
 	static BOOL SetText(tVariant* paParams, const long lSizeArray);
 	static BOOL SetWindowState(tVariant* paParams, const long lSizeArray);
 	static BOOL SetWindowSize(tVariant* paParams, const long lSizeArray);
